+++ lib/marhan_cli/command.rb
@@ -13,6 +13,8 @@ module MarhanCli
     def execute(proc)
       begin
         say ""
+        say "Starting command...", :green
+        say ""
         proc.call
         say ""
         say "End of command", :green
@@ -31,7 +33,6 @@ module MarhanCli
       unless File.exists?(config_file_path)
         raise "Stop processing! Command needs the configuration file '#{CONFIG_FILE}' in you're home directory."
       end
-      #say "I will use the '#{CONFIG_FILE}' configuration file in you're home directory.", :cyan
       config = Ambience.create(config_file_path)
       config.to_mash
     end